+++ super_test_engine/lib/super_test_engine/copy_app_templates/controllers/sinks_controller.rb
@@ -1,9 +1,7 @@
-# frozen_string_literal: true
-
 class Admin::SinksController < AdminController
-  class Controls < AdminControls
-    def model
-      Sink
-    end
+  private
+
+  def model
+    Sink
   end
 end