+++ test/dummy/rails52_sprockets4/app/controllers/admin_controller.rb
@@ -1,10 +1,2 @@
 class AdminController < Super::ApplicationController
-  class AdminControls < Super::Controls
-  end
-
-  private
-
-  def new_controls
-    self.class::Controls.new
-  end
 end