+++ test/dummy/rails60_webpacker4/app/controllers/admin/sinks_controller.rb
@@ -1,7 +1,7 @@
 class Admin::SinksController < AdminController
-  class Controls < AdminControls
-    def model
-      Sink
-    end
+  private
+
+  def model
+    Sink
   end
 end